Modularize code and add specialized H for 1 arm.
